# web/public/openapi-integration.yaml
...
components:
    parameters:
...
        orgId:
            description: Organisation id
            in: path
            name: id
            required: true
            schema:
                type: string
...
    responses:
        Error:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
            description: Error response
    schemas:
...
        Entitlement:
            properties:
                description:
                    type: string
                id:
                    type: string
                key:
                    type: string
                scope:
                    enum:
                        - platform
                        - product
                    type: string
            type: object
        Error:
            properties:
                code:
                    description: Machine-readable code (e.g. not_found, validation_error, unauthorized, forbidden, rate_limited)
                    type: string
                message:
                    type: string
            required:
                - code
                - message
            type: object
            x-go-name: APIError
        ErrorResponse:
            properties:
                error:
                    $ref: '#/components/schemas/Error'
            required:
                - error
            type: object
...
        Subscription:
            properties:
                current_period_end:
                    format: date-time
                    nullable: true
                    type: string
                id:
                    type: string
                organisation_id:
                    type: string
                plan_id:
                    type: string
                processor:
                    type: string
                status:
                    enum:
                        - trialing
                        - active
                        - past_due
                        - canceled
                    type: string
                subscription_ref:
                    type: string
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            description: Session token (kyc_sess_…) or API key / service token
            scheme: bearer
            type: http
info:
    description: |-
        JSON API for merchant backends integrating with KYC (app users, attributes, product plans, entitlements).

        Authenticate with an **organisation API key**: `Authorization: Bearer kyc_…` (create keys under Platform → API keys).
        Keys may be scoped to RBAC permissions; empty scopes grant full organisation access. The organisation must have the `api_access` entitlement.

        Operator UI routes (Google OAuth, members, roles, permissions, email/automation config, platform admin) are documented separately under **Operator API**.
    title: KYC Integration API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /v1/entitlements:
        get:
            operationId: listEntitlements
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    items:
                                        items:
                                            $ref: '#/components/schemas/Entitlement'
                                        type: array
                                type: object
                    description: Entitlement list
                default:
                    $ref: '#/components/responses/Error'
            summary: List global entitlement catalog
            tags:
                - Billing
...
    /v1/organisations/{id}/subscription:
        get:
            description: Requires billing:read.
            operationId: getSubscription
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Subscription'
                    description: Subscription
                default:
                    $ref: '#/components/responses/Error'
            summary: Get organisation subscription
            tags:
                - Billing
        parameters:
            - $ref: '#/components/parameters/orgId'
...
security:
    - bearerAuth: []
servers:
    - description: Relative to the web app proxy (paths include `/v1`)
      url: /
...